♻️ Refactor: 아이돌 검색 페이지 구조 분리 (#61)
